telegram_bot/requirements.txt
python-telegram-bot[ext]>=20.0  # Usamos la versión 20+ que es async
httpx[http2] # Dependencia a menudo necesaria
requests